Final Project for LING 499J
In this project, I made a change to a model that used Vectors to represent word meaning by using a corpus of words and calculating how often each word appeared within a pre-determined distance of another word.  The original model did not use parts of speech (POS) in generating the co-occurrence counts, which I believe to be a shortcoming of said model. The new model retains the POS tags in the corpus so that it can better capture words that have different parts of speech and/or meanings while their written form remains the same.  I evaluated the old model without POS tags to the new model with POS tags by comparing the similarity judgements each model produced to actual human similarity judgements in order to determine which model produces judgments that are closer to humans.